+++ packages/backend-tasks/src/tasks/TaskWorker.ts
@@ -145,6 +145,7 @@ export class TaskWorker {
       await this.fn(taskAbortController.signal);
       taskAbortController.abort(); // releases resources
     } catch (e) {
+      this.logger.error(e);
       await this.tryReleaseTask(ticket, taskSettings);
       return { result: 'failed' };
     } finally {
@@ -174,8 +175,9 @@ export class TaskWorker {
     } else if (isCron) {
       const time = new CronTime(settings.cadence)
         .sendAt()
-        .add({ seconds: -1 }) // immediately, if "* * * * * *"
-        .toISOString();
+        .minus({ seconds: 1 }) // immediately, if "* * * * * *"
+        .toUTC()
+        .toISO();
       startAt = this.knex.client.config.client.includes('sqlite3')
         ? this.knex.raw('datetime(?)', [time])
         : this.knex.raw(`?`, [time]);
@@ -278,7 +280,7 @@ export class TaskWorker {
 
     let nextRun: Knex.Raw;
     if (isCron) {
-      const time = new CronTime(settings.cadence).sendAt().toISOString();
+      const time = new CronTime(settings.cadence).sendAt().toUTC().toISO();
       this.logger.debug(`task: ${this.taskId} will next occur around ${time}`);
       nextRun = this.knex.client.config.client.includes('sqlite3')
         ? this.knex.raw('datetime(?)', [time])