+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/filesystem/rename.ts
@@ -66,6 +66,7 @@ export const createFilesystemRenameAction = () => {
         },
       },
     },
+    supportsDryRun: true,
     async handler(ctx) {
       if (!Array.isArray(ctx.input?.files)) {
         throw new InputError('files must be an Array');