+++ packages/backend-common/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/backend-common
 
+## 0.13.6-next.1
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/config-loader@1.1.2-next.0
+  - @backstage/integration@1.2.1-next.1
+
 ## 0.13.6-next.0
 
 ### Patch Changes